+++ src/main/java/com/telepathicgrunt/the_bumblezone/capabilities/IEntityPosAndDim.java
@@ -9,11 +9,7 @@ public interface IEntityPosAndDim {
 	//what methods the capability will have and what the capability is
 	
 	void setNonBZDim(ResourceLocation incomingDim);
-	void setTeleporting(boolean teleporting);
-
 	ResourceLocation getNonBZDim();
-	boolean getTeleporting();
-
 	void setNonBZPos(Vector3d incomingPos);
 	Vector3d getNonBZPos();
 	void setNonBZPitch(float incomingPitch);